Add more tests, and split the compilation of graph.cpp in 3 files to speed up the compilation.
