+++ src/python_interface/rcspp/graph.cpp
@@ -1,54 +1,21 @@
 // Copyright (c) 2025 Laboratory for Combinatorial Optimization in Real-time Environment.
 // All rights reserved.
 
-#define PYBIND11_USE_SMART_HOLDER_AS_DEFAULT
-#include "rcspp/graph/graph.hpp"
-
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
-
-#include <atomic>
+// graph_impl.hpp defines PYBIND11_USE_SMART_HOLDER_AS_DEFAULT before the pybind11 includes.
 #include <csignal>
-#include <limits>
-#include <memory>
-#include <tuple>
 
-#include "rcspp/algorithm/greedy.hpp"
-#include "rcspp/algorithm/pulling_dominance_algorithm.hpp"
-#include "rcspp/algorithm/simple_dominance_algorithm.hpp"
-#include "rcspp/resource/concrete/container_resource.hpp"
-#include "rcspp/resource/concrete/numerical_resource.hpp"
-#include "rcspp/resource/resource_graph.hpp"
-#include "resource_types.hpp"
+#include "graph_impl.hpp"
 
-namespace py = pybind11;
-
-using namespace rcspp;
-
-// ─── SIGINT interruptibility for Python ───────────────────────────────────────
-// The handler is installed ONCE at module import time (from the main thread, via
-// init_sigint_handler() called in PYBIND11_MODULE).
-//
-// Any long-running C++ binding that releases the GIL follows this pattern:
-//
-//   { py::gil_scoped_release _; result = dispatch_algorithm<>(...); }
-//   // dispatch_algorithm wires g_py_interrupted + manages g_active_calls
-//   if (g_py_interrupted.exchange(false)) { raise KeyboardInterrupt; }
-//
-// The g_active_calls counter tells the handler whether a C++ call is in
-// progress.  When it is, the handler sets the flag but does NOT forward to
-// Python's trampoline — the binding raises KeyboardInterrupt itself once the
-// GIL is re-acquired.  When no call is active, the handler forwards so that
-// ordinary Ctrl-C still raises KeyboardInterrupt in Python code.
-
-static std::atomic<bool> g_py_interrupted{false};
-static std::atomic<int> g_active_calls{0};
+// ─── SIGINT state ─────────────────────────────────────────────────────────────
+// Declared extern in graph_impl.hpp; defined here (one TU owns the storage).
+std::atomic<bool> g_py_interrupted{false};
+std::atomic<int> g_active_calls{0};
 static struct sigaction g_old_sigint_sa{};
 
 static void py_sigint_handler(int sig) {
     g_py_interrupted.store(true, std::memory_order_relaxed);
-    // Forward to Python's handler only when no C++ call holds the guard;
-    // otherwise we raise KeyboardInterrupt ourselves after the call returns.
+    // Forward to Python's handler only when no C++ call is active; otherwise
+    // the binding raises KeyboardInterrupt itself after re-acquiring the GIL.
     if (g_active_calls.load(std::memory_order_relaxed) == 0) {
         auto* h = g_old_sigint_sa.sa_handler;
         if (h != nullptr && h != SIG_DFL && h != SIG_IGN) {
@@ -66,359 +33,11 @@ void init_sigint_handler() {
     sigaction(SIGINT, &new_sa, &g_old_sigint_sa);
 }
 
-// Releases the GIL, calls f(), re-acquires the GIL, then raises KeyboardInterrupt
-// if g_py_interrupted was set during the call.  Handles both void and non-void callables.
-// Usage: return run_interruptible([&]{…});   (void callable: just call without return)
-template <typename F>
-auto run_interruptible(F&& f) {
-    auto check = [] {
-        if (g_py_interrupted.exchange(false, std::memory_order_relaxed)) {
-            PyErr_SetNone(PyExc_KeyboardInterrupt);
-            throw py::error_already_set();
-        }
-    };
-    if constexpr (std::is_void_v<std::invoke_result_t<F>>) {
-        {
-            py::gil_scoped_release release;
-            std::forward<F>(f)();
-        }
-        check();
-    } else {
-        auto result = [&] {
-            py::gil_scoped_release release;
-            return std::forward<F>(f)();
-        }();
-        check();
-        return result;
-    }
-}
-
 // ─── Concrete type aliases ────────────────────────────────────────────────────
 
 using RealRC = ResourceComposition<RealResource>;
 using RealGraph = Graph<RealRC>;
 using RealRG = ResourceGraph<RealResource>;
-
-// ─── Algorithm dispatch table ─────────────────────────────────────────────────
-// Add new algorithms here only; dispatch is driven automatically.
-
-enum class SolverAlgorithm { Simple, Pulling, Greedy };
-
-// Associates one SolverAlgorithm enum value with its algorithm template.
-template <SolverAlgorithm E, template <typename> class Algo>
-struct AlgoEntry {
-        static constexpr SolverAlgorithm value = E;
-        template <typename RG, typename CostRC>
-        static std::vector<Solution> run(RG& rg, double ub, AlgorithmParams p, bool pre, int ci) {
-            return rg.template solve<Algo, CostRC>(ub, p, pre, ci);
-        }
-};
-
-// Single source of truth for the enum → algorithm template mapping.
-using AlgorithmTable = std::tuple<AlgoEntry<SolverAlgorithm::Simple, SimpleDominanceAlgorithm>,
-                                  AlgoEntry<SolverAlgorithm::Pulling, PullingDominanceAlgorithm>,
-                                  AlgoEntry<SolverAlgorithm::Greedy, GreedyAlgorithm>>;
-
-template <typename RG, typename CostRC, typename... Entries>
-std::vector<Solution> dispatch_algorithm_impl(SolverAlgorithm alg, RG& rg, double ub,
-                                              AlgorithmParams p, bool pre, int ci,
-                                              std::tuple<Entries...>* /*tag*/) {
-    std::vector<Solution> result;
-    [[maybe_unused]] bool matched =
-        ((Entries::value == alg
-              ? (result = Entries::template run<RG, CostRC>(rg, ub, p, pre, ci), true)
-              : false) ||
-         ...);
-    return result;
-}
-
-template <typename RG, typename CostRC>
-std::vector<Solution> dispatch_algorithm(SolverAlgorithm alg, RG& rg, double ub, AlgorithmParams p,
-                                         bool pre, int ci) {
-    // Wire up the global interrupt flag and track that a C++ call is active so
-    // the signal handler knows not to forward to Python's handler (which would
-    // cause a double KeyboardInterrupt in the calling thread).
-    p.interrupted = &g_py_interrupted;
-    g_py_interrupted.store(false, std::memory_order_relaxed);
-    g_active_calls.fetch_add(1, std::memory_order_relaxed);
-    auto result = dispatch_algorithm_impl<RG, CostRC>(alg,
-                                                      rg,
-                                                      ub,
-                                                      p,
-                                                      pre,
-                                                      ci,
-                                                      static_cast<AlgorithmTable*>(nullptr));
-    g_active_calls.fetch_sub(1, std::memory_order_relaxed);
-    return result;
-}
-
-// ─── Resource type → pybind11 method name ────────────────────────────────────
-// Generated for all resource types via X-macro.
-
-template <typename T>
-constexpr const char* add_resource_method_name();
-
-#define GEN_ADD_RESOURCE_NAME(name, scalar, RT)            \
-    template <>                                            \
-    constexpr const char* add_resource_method_name<RT>() { \
-        return "add_" #name "_resource";                   \
-    }
-RCSPP_ALL_RESOURCES(GEN_ADD_RESOURCE_NAME)
-#undef GEN_ADD_RESOURCE_NAME
-
-// ─── Python type name per resource type ──────────────────────────────────────
-// Returns the prefix used in Python class names ("real", "int", "uint_bitset", …).
-
-template <typename T>
-constexpr const char* py_type_name();
-
-#define GEN_PY_TYPE_NAME(name, scalar, RT)     \
-    template <>                                \
-    constexpr const char* py_type_name<RT>() { \
-        return #name;                          \
-    }
-RCSPP_ALL_RESOURCES(GEN_PY_TYPE_NAME)
-#undef GEN_PY_TYPE_NAME
-
-// ─── CostRC auto-selection ────────────────────────────────────────────────────
-// Picks the first numerical resource in the pack; falls back to RealResource sentinel.
-
-template <typename... RTs>
-struct SelectCostRC {
-        using type = RealResource;
-};
-
-template <typename RT, typename... RTs>
-struct SelectCostRC<RT, RTs...> {
-        using type = std::conditional_t<is_numerical_resource_v<RT>, RT,
-                                        typename SelectCostRC<RTs...>::type>;
-};
-
-// ─── Forward declaration (definition follows bind_resource_graph_block) ───────
-template <typename CostRC, typename... RTs>
-void bind_mixed_rg(py::module_& m, const char* name);
-
-// ─── Auto class-name builder and binder ──────────────────────────────────────
-
-template <typename... RTs>
-std::string auto_mix_name() {
-    std::vector<const char*> names = {py_type_name<RTs>()...};
-    std::string s;
-    for (size_t i = 0; i < names.size(); ++i) {
-        if (i > 0) {
-            s += '_';
-        }
-        s += names[i];
-    }
-    return s;
-}
-
-template <typename... RTs>
-void bind_auto_rg(py::module_& m) {
-    using CostRC = typename SelectCostRC<RTs...>::type;
-    bind_mixed_rg<CostRC, RTs...>(m, auto_mix_name<RTs...>().c_str());
-}
-
-#define BIND_MIX(...) bind_auto_rg<__VA_ARGS__>(m)
-
-// ─── Helper: bind common Graph base methods ───────────────────────────────────
-
-template <typename G>
-py::class_<G>& bind_graph_methods(py::class_<G>& c) {
-    return c.def("get_node", &G::get_node, py::arg("id"), py::return_value_policy::reference)
-        .def("get_arc", &G::get_arc, py::arg("id"), py::return_value_policy::reference)
-        .def("node_ids", &G::get_node_ids)
-        .def("arc_ids", &G::get_arc_ids)
-        .def("source_node_ids", &G::get_source_node_ids)
-        .def("sink_node_ids", &G::get_sink_node_ids)
-        .def("number_of_nodes", &G::get_number_of_nodes)
-        .def("number_of_arcs", &G::get_number_of_arcs)
-        .def("is_source", &G::is_source, py::arg("node_id"))
-        .def("is_sink", &G::is_sink, py::arg("node_id"))
-        .def("to_string", &G::to_string, py::arg("print_arcs") = false)
-        .def("__str__", [](const G& g) { return g.to_string(); })
-        .def("__repr__", [](const G& g) { return g.to_string(); });
-}
-
-// ─── Helper: bind common ResourceGraph methods (algorithm dispatch included) ──
-
-template <typename RG, typename RC, typename CostRC = RealResource>
-py::class_<RG, Graph<RC>>& bind_rg_methods(py::class_<RG, Graph<RC>>& c) {
-    using N = Node<RC>;
-    constexpr double INF = std::numeric_limits<double>::infinity();
-
-    return c
-        .def("add_node",
-             static_cast<N& (RG::*)(size_t, bool, bool)>(&RG::add_node),
-             py::arg("id"),
-             py::arg("source") = false,
-             py::arg("sink") = false,
-             py::return_value_policy::reference)
-        .def("get_resource_factory", &RG::get_resource_factory, py::return_value_policy::reference)
-        .def(
-            "solve",
-            [](RG& rg, SolverAlgorithm alg, double ub, AlgorithmParams p, bool pre, int ci)
-                -> std::vector<Solution> {
-                return run_interruptible(
-                    [&] { return dispatch_algorithm<RG, CostRC>(alg, rg, ub, p, pre, ci); });
-            },
-            py::arg("algorithm") = SolverAlgorithm::Simple,
-            py::arg("upper_bound") = INF,
-            py::arg("params") = AlgorithmParams{},
-            py::arg("preprocess") = true,
-            py::arg("cost_index") = 0)
-        .def("process_feasibility",
-             [](RG& rg) { run_interruptible([&] { rg.process_feasibility(); }); })
-        .def(
-            "is_connected",
-            [](RG& rg, size_t o, size_t d) {
-                return run_interruptible([&] { return rg.is_connected(o, d); });
-            },
-            py::arg("origin_node_id"),
-            py::arg("destination_node_id"));
-}
-
-// ─── Helper: bind one add_resource method for a specific resource type ────────
-
-template <typename RG, typename RC, typename ResourceType>
-void bind_add_resource(py::class_<RG, Graph<RC>>& rg) {
-    rg.def(add_resource_method_name<ResourceType>(),
-           static_cast<void (RG::*)(std::unique_ptr<ExtensionFunction<ResourceType>>,
-                                    std::unique_ptr<FeasibilityFunction<ResourceType>>,
-                                    std::unique_ptr<CostFunction<ResourceType>>,
-                                    std::unique_ptr<DominanceFunction<ResourceType>>)>(
-               &RG::template add_resource<ResourceType>),
-           py::arg("extension_function"),
-           py::arg("feasibility_function"),
-           py::arg("cost_function"),
-           py::arg("dominance_function"));
-}
-
-// ─── Helper: bind resource-specific methods (add_resource*, add_arc, ──────────
-//             update_arc, and update_reduced_costs when RealResource is present)
-
-template <typename RG, typename RC, typename... ResourceTypes>
-void bind_resource_graph_impl(py::class_<RG, Graph<RC>>& rg) {
-    using AddArcTuple = std::tuple<std::vector<ResourceInitializerTypeTuple_t<ResourceTypes>>...>;
-
-    // add_<type>_resource — one per resource type via fold
-    (bind_add_resource<RG, RC, ResourceTypes>(rg), ...);
-
-    rg.def("add_arc",
-           static_cast<Arc<RC>& (RG::*)(const AddArcTuple&,
-                                        size_t,
-                                        size_t,
-                                        std::optional<size_t>,
-                                        double,
-                                        std::vector<Row>)>(&RG::add_arc),
-           py::arg("resource_consumption"),
-           py::arg("origin_node_id"),
-           py::arg("destination_node_id"),
-           py::arg("id") = std::nullopt,
-           py::arg("cost") = 0.0,
-           py::arg("dual_rows") = std::vector<Row>{},
-           py::return_value_policy::reference);
-
-    rg.def("update_arc",
-           static_cast<void (RG::*)(Arc<RC>*, const AddArcTuple&, std::optional<double>)>(
-               &RG::update_arc),
-           py::arg("arc"),
-           py::arg("resource_consumption"),
-           py::arg("cost") = std::nullopt);
-
-    // update_reduced_costs is only meaningful when RealResource is present
-    if constexpr ((std::is_same_v<ResourceTypes, RealResource> || ...)) {
-        rg.def(
-            "update_reduced_costs",
-            [](RG& rg, const std::vector<double>& duals, size_t cost_index) {
-                run_interruptible(
-                    [&] { rg.template update_reduced_costs<RealResource>(duals, cost_index); });
-            },
-            py::arg("duals"),
-            py::arg("cost_index") = 0);
-    }
-}
-
-// ─── Helper: build the full block for a set of resource types ─────────────────
-// Registers Node, Arc, plain Graph, and ResourceGraph under the given Python names.
-
-template <typename RG, typename RC, typename CostRC, typename... ResourceTypes>
-void bind_resource_graph_block(py::module_& m, const char* rg_name, const char* graph_name,
-                               const char* node_name, const char* arc_name) {
-    py::class_<Node<RC>>(m, node_name)
-        .def_readonly("id", &Node<RC>::id)
-        .def_readonly("source", &Node<RC>::source)
-        .def_readonly("sink", &Node<RC>::sink)
-        .def("__str__", &Node<RC>::to_string)
-        .def("__repr__", &Node<RC>::to_string);
-
-    py::class_<Arc<RC>>(m, arc_name)
-        .def_readonly("id", &Arc<RC>::id)
-        .def(
-            "origin",
-            [](const Arc<RC>& a) -> Node<RC>* { return a.origin; },
-            py::return_value_policy::reference)
-        .def(
-            "destination",
-            [](const Arc<RC>& a) -> Node<RC>* { return a.destination; },
-            py::return_value_policy::reference)
-        .def_readwrite("cost", &Arc<RC>::cost)
-        .def_readwrite("dual_rows", &Arc<RC>::dual_rows)
-        .def("__str__", &Arc<RC>::to_string)
-        .def("__repr__", &Arc<RC>::to_string);
-
-    {
-        py::class_<Graph<RC>> g(m, graph_name);
-        bind_graph_methods(g);
-    }
-
-    py::class_<RG, Graph<RC>> rg(m, rg_name);
-    bind_rg_methods<RG, RC, CostRC>(rg);
-    rg.def(py::init<>());
-    bind_resource_graph_impl<RG, RC, ResourceTypes...>(rg);
-}
-
-// ─── Helper: bind a mixed-resource graph under a derived Python name ─────────
-// CostRC   — the resource type used for cost-based preprocessing (must be numerical).
-// RTs...   — the full ordered resource-type pack (same as template arguments to ResourceGraph).
-// name     — the middle part of the Python class name; the binding registers
-//            "_<name>_resource_graph", "_<name>_graph", "_<name>_node", "_<name>_arc".
-
-template <typename CostRC, typename... RTs>
-void bind_mixed_rg(py::module_& m, const char* name) {
-    using RC = ResourceComposition<RTs...>;
-    using RG = ResourceGraph<RTs...>;
-    std::string rg = std::string("_") + name + "_resource_graph";
-    std::string g = std::string("_") + name + "_graph";
-    std::string n = std::string("_") + name + "_node";
-    std::string a = std::string("_") + name + "_arc";
-    bind_resource_graph_block<RG, RC, CostRC, RTs...>(m,
-                                                      rg.c_str(),
-                                                      g.c_str(),
-                                                      n.c_str(),
-                                                      a.c_str());
-}
-
-// ─── Macros: bind single-resource graph blocks ───────────────────────────────
-// Numerical: CostRC = RT (real/int/uint have a meaningful cost resource).
-// Container: CostRC = RealResource (containers don't have a numeric cost resource;
-//            RealResource satisfies requires, and the "not in pack" guard skips preprocessing).
-
-#define BIND_SINGLE_NUMERICAL_RG(name, scalar, RT)                                 \
-    bind_resource_graph_block<ResourceGraph<RT>, ResourceComposition<RT>, RT, RT>( \
-        m,                                                                         \
-        "_" #name "_resource_graph",                                               \
-        "_" #name "_graph",                                                        \
-        "_" #name "_node",                                                         \
-        "_" #name "_arc");
-
-#define BIND_SINGLE_CONTAINER_RG(name, scalar, RT)                                           \
-    bind_resource_graph_block<ResourceGraph<RT>, ResourceComposition<RT>, RealResource, RT>( \
-        m,                                                                                   \
-        "_" #name "_resource_graph",                                                         \
-        "_" #name "_graph",                                                                  \
-        "_" #name "_node",                                                                   \
-        "_" #name "_arc");
 
 // ─── init_graph ───────────────────────────────────────────────────────────────
 
@@ -538,39 +157,14 @@ void init_graph(py::module_& m) {
     }
 
     // ══════════════════════════════════════════════════════════════════════════
-    // Single-resource graphs — generated via X-macro -> RealResource is generated by default
+    // Single-resource graphs (generated via X-macro; RealResource bound above)
     // ══════════════════════════════════════════════════════════════════════════
     BIND_SINGLE_NUMERICAL_RG(int, int, IntResource)
 
     // ══════════════════════════════════════════════════════════════════════════
-    // Mixed graphs — all non-trivial subsets of (real, int, real_set, int_set,
-    // uint_bitset), grouped by size.  CostRC = first numerical type in the pack,
-    // or RealResource as a sentinel for all-container combos (preprocessing is
-    // skipped by the "not in pack" if-constexpr guard in resource_graph.hpp).
-    //
-    // The Python class name is "_<t1>_<t2>_..._resource_graph" where each ti is
-    // the C++ prefix (real, int, real_set, int_set, uint_bitset).
+    // Mixed-resource graphs — split across graph_mix2.cpp and graph_mix3.cpp
+    // for faster parallel compilation.
     // ══════════════════════════════════════════════════════════════════════════
-
-    // clang-format off
-
-    // ── Pairs (2-type) ────────────────────────────────────────────────────────
-    BIND_MIX(RealResource, IntResource);
-    BIND_MIX(RealResource, RealSetResource);
-    BIND_MIX(RealResource, IntSetResource);
-    BIND_MIX(RealResource, UIntBitsetResource);
-    BIND_MIX(IntResource,  RealSetResource);
-    BIND_MIX(IntResource,  IntSetResource);
-    BIND_MIX(IntResource,  UIntBitsetResource);
-
-    // ── Triples (3-type) ──────────────────────────────────────────────────────
-    BIND_MIX(RealResource, IntResource,     RealSetResource);
-    BIND_MIX(RealResource, IntResource,     IntSetResource);
-    BIND_MIX(RealResource, IntSetResource,  UIntBitsetResource);
-
-    // ── Universal (all 5 types) ───────────────────────────────────────────────
-    BIND_MIX(RealResource, IntResource, RealSetResource, IntSetResource, UIntBitsetResource);
-    // clang-format on
-
-#undef BIND_MIX
+    init_graph_mix2(m);
+    init_graph_mix3(m);
 }
